% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver Queue System Simulation}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A numeric vector of customer arrival times.}

\item{ServiceTimes}{A numeric vector of service times for each customer.}

\item{NumServers}{An integer specifying the number of servers available (default is 1).}
}
\value{
A tibble containing the following columns:
\describe{
\item{Arrivals}{The time when each customer arrived.}
\item{ServiceBegins}{The time when each customer's service began.}
\item{ChosenServer}{The server that served each customer.}
\item{ServiceEnds}{The time when each customer's service ended.}
}
}
\description{
Simulates a multi-server queue system where customers are served based on their arrival and service times.
This function calculates the service start and end times for each customer based on the availability of servers.
}
\examples{
set.seed(2048)
arrival_time <- cumsum(rexp(100, 1/60))
service_time <- rexp(length(arrival_time), 1/150) + 20
Multiserver(arrival_time, service_time, 3)
}
